third_party/blink/web_tests/virtual/mutation-events-enabled/fast/events/mutation-events-disabled-expected.txt

CONSOLE WARNING: Listener added for a 'DOMSubtreeModified' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMSubtreeModified' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeInserted' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeInserted' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeRemoved' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeRemoved' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeRemovedFromDocument' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeRemovedFromDocument' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeInsertedIntoDocument' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeInsertedIntoDocument' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMCharacterDataModified' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMCharacterDataModified' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMNodeRemoved event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMNodeRemoved event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMNodeRemovedFromDocument event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMSubtreeModified event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMNodeInserted event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMNodeInserted event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMNodeInsertedIntoDocument event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMSubtreeModified event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMSubtreeModified event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMSubtreeModified event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMCharacterDataModified event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMCharacterDataModified event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMSubtreeModified event Reached unreachable code
CONSOLE ERROR: Uncaught Error: assert_unreached: Received DOMSubtreeModified event Reached unreachable code
CONSOLE WARNING: Listener added for a 'DOMSubtreeModified' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeInserted' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeRemoved' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeRemovedFromDocument' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeInsertedIntoDocument' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMCharacterDataModified' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught Error: assert_unreached: Received DOMSubtreeModified event Reached unreachable code
Harness: the test ran to completion.
